KIEKER-1916 marked deprecated classes and moved new stages in the right location.
