+++ kieker-tools/log-replayer/src/kieker/tools/log/replayer/FilesystemLogReplayer.java
@@ -26,11 +26,13 @@ import kieker.common.configuration.Configuration;
 /**
  * An implementation of the {@link AbstractLogReplayer}, using the {@link FSReader} to replay {@link kieker.common.record.IMonitoringRecord}s from a list of file
  * system monitoring logs.
- * 
+ *
  * @author Andre van Hoorn
- * 
+ *
  * @since 0.95a
+ * @deprecated 1.16
  */
+@Deprecated
 public class FilesystemLogReplayer extends AbstractLogReplayer {
 	// private static final Log LOG = LogFactory.getLog(FilesystemLogReplayer.class);
 
@@ -38,7 +40,7 @@ public class FilesystemLogReplayer extends AbstractLogReplayer {
 
 	/**
 	 * Creates a new instance of this class using the given parameters.
-	 * 
+	 *
 	 * @param monitoringConfigurationFile
 	 *            The name of the monitoring configuration file.
 	 * @param realtimeMode
