+++ kieker-tools/log-replayer/src/kieker/tools/log/replayer/JMSLogReplayer.java
@@ -23,11 +23,13 @@ import kieker.common.configuration.Configuration;
 
 /**
  * An implementation of the {@link AbstractLogReplayer}, using the {@link JmsReader} to replay {@link kieker.common.record.IMonitoringRecord}s from a JMS queue.
- * 
+ *
  * @author Andre van Hoorn
- * 
+ *
  * @since 1.3
+ * @deprecated since 1.16
  */
+@Deprecated
 public class JMSLogReplayer extends AbstractLogReplayer {
 
 	private final String jmsProviderUrl;
@@ -36,7 +38,7 @@ public class JMSLogReplayer extends AbstractLogReplayer {
 
 	/**
 	 * Creates a new JMS log replayer.
-	 * 
+	 *
 	 * @param jmsProviderUrl
 	 *            = for instance "tcp://127.0.0.1:3035/"
 	 * @param jmsDestination
