+++ kieker-tools/log-replayer/src/kieker/tools/log/replayer/ReplayerMain.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  ***************************************************************************/
-package kieker.tools.log.replayer.teetime;
+package kieker.tools.log.replayer;
 
 import java.io.File;
