+++ kieker-tools/log-replayer/src/kieker/tools/log/replayer/package-info.java
@@ -14,4 +14,3 @@
  * limitations under the License.
  ***************************************************************************/
 package kieker.tools.log.replayer;
-
